requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anthropic==0.47.2
anyio==4.9.0
appnope==0.1.4
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
backcall==0.2.0
beautifulsoup4==4.13.4
bleach==6.2.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.1
control==0.10.1
croissant==0.4
cvxopt==1.3.2
cycler==0.12.1
datasets==3.6.0
decorator==5.2.1
defusedxml==0.7.1
dill==0.3.8
distro==1.9.0
docopt==0.6.2
docstring_parser==0.16
exceptiongroup==1.2.2
executing==2.2.0
fastjsonschema==2.21.1
ffmpeg==1.4
filelock==3.18.0
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2025.3.0
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.25.0rc0
google-api-python-client==2.169.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-genai==1.13.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.31.2
idna==3.10
imageio==2.37.0
instructor==1.7.8
ipython==8.12.3
jedi==0.19.2
Jinja2==3.1.6
jiter==0.8.2
joblib==1.4.2
jsonref==1.1.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
kiwisolver==1.4.8
lazy_loader==0.4
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matlabengine==24.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.3
mpmath==1.3.0
multidict==6.2.0
multiprocess==0.70.16
narwhals==1.34.1
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
networkx==3.4.2
nltk==3.9.1
numpy==2.2.4
openai==1.68.2
opencv-python==4.11.0.86
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.1.0
pipreqs==0.5.0
platformdirs==4.3.8
prompt_toolkit==3.0.51
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.10.6
pydantic_core==2.27.2
pydeck==0.9.1
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.4.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rpds-py==0.24.0
rsa==4.9.1
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
stack-data==0.6.3
streamlit==1.44.1
sympy==1.13.3
tenacity==9.1.2
threadpoolctl==3.6.0
tifffile==2025.3.30
tiktoken==0.9.0
tinycss2==1.4.0
toml==0.10.2
torch==2.7.0
torchvision==0.22.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typer==0.15.2
typing_extensions==4.12.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.3.0
wcwidth==0.2.13
webencodings==0.5.1
websockets==15.0.1
wordcloud==1.9.4
xxhash==3.5.0
yarg==0.1.9
yarl==1.18.3
